Fix how protoc proto_path is generated to add derived directories
When this plugin is run under m2eclipse, it inherits the classpath
of other m2eclipse projects which might contain protocol buffer files.
Instead of not using them, we now add them as the proto_path.
It is then possible to spread proto files in several maven projects
as long as dependencies are correctly set.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>